Refactored erorrs
