+++ app/api/v1/exceptions/websocket/internal_server_error.py
@@ -1,5 +1,6 @@
+from app.api.v1.exceptions.websocket import websocket_status
 from app.api.v1.exceptions.websocket.websocket_error import WebSocketError
 
 
 class InternalServerError(WebSocketError):
-    status_code = 4100
+    status_code = websocket_status.WS_4100_INTERNAL_ERROR
