+++ app/api/v1/exceptions/websocket/invalid_packet_data.py
@@ -1,5 +1,6 @@
+from app.api.v1.exceptions.websocket import websocket_status
 from app.api.v1.exceptions.websocket.websocket_error import WebSocketError
 
 
 class InvalidPacketDataError(WebSocketError):
-    status_code = 4000
+    status_code = websocket_status.WS_4000_BAD_REQUEST
