+++ app/api/v1/exceptions/websocket/not_authenticated_address.py
@@ -1,5 +1,6 @@
+from app.api.v1.exceptions.websocket import websocket_status
 from app.api.v1.exceptions.websocket.websocket_error import WebSocketError
 
 
 class NotAuthenticatedAddressError(WebSocketError):
-    status_code = 4002
+    status_code = websocket_status.WS_4001_UNAUTHORIZED
