+++ app/api/v1/exceptions/websocket/websocket_error.py
@@ -1,2 +1,5 @@
+from app.api.v1.exceptions.websocket import websocket_status
+
+
 class WebSocketError(Exception):
-    status_code = 4000
+    status_code = websocket_status.WS_4000_BAD_REQUEST
