+++ app/api/v1/exceptions/websocket/websocket_status.py
@@ -1,0 +1,4 @@
+WS_4000_BAD_REQUEST = 4000
+WS_4001_UNAUTHORIZED = 4001
+
+WS_4100_INTERNAL_ERROR = 4100
