+++ app/api/v1/router.py
@@ -1,8 +1,8 @@
 from fastapi import APIRouter
 
-from app.api.v1.routes.auth import auth_router
-from app.api.v1.routes.games import games_router
-from app.api.v1.routes.games_packets import games_packets_router
+from app.api.v1.routes.http.auth import auth_router
+from app.api.v1.routes.http.games import games_router
+from app.api.v1.routes.websocket.games import games_packets_router
 
 v1_router: APIRouter = APIRouter(prefix="/v1")
